Manage Namespace and support OpenShift
csi-driver-smb has to be updated to v1.7.0 due to an upstream bugfix
which added the `serviceAccountName` to the node DaemonSet.

The Helm Chart can create the ServiceAccounts and we therefore don't
have to manage them in this component.

The Pods of this CSIDriver run in privileged mode and therefore require
the privileged SCC on OpenShift.

Closes #22